# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library API",
        "description": "API documentation for the library project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/books/{book_id}/favorites": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "add book to favorities lists",
                "operationId": "09159c036cfedfc3cd49f8718d0c0c3b",
                "parameters": [
                    {
                        "name": "book_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "book has been add it to the favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "add book to favorities lists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "remove the book from favorites list",
                "operationId": "2450986e3ac0e9c9619b52043aafc6e2",
                "parameters": [
                    {
                        "name": "book_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "book has been removed from the favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "book removed from favorites"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/books/{book_id}/saves": {
            "post": {
                "tags": [
                    "Saves"
                ],
                "summary": "create a new review for book",
                "operationId": "5769dbac24ecffb7feedcbdfcb0e73d2",
                "parameters": [
                    {
                        "name": "book_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "book has been add it to the saves",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "book add it to saves"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Saves"
                ],
                "summary": "remove the book from saves list",
                "operationId": "59afa28faedcc4357fe94c2794fb1cd8",
                "parameters": [
                    {
                        "name": "book_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "book has been removed from the saves",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "book removed from saves"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/{provider}/redirect": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get OAuth redirect URL",
                "operationId": "3b3c6afe252d49a19fc747b57b1a7d9c",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "google"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Redirect URL",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "redirect_url": {
                                            "type": "string",
                                            "example": "https://accounts.google.com/..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid provider"
                    }
                }
            }
        },
        "/auth/{provider}/callback": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "OAuth callback",
                "operationId": "77d9e10d85fca305876c149607afd051",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "google",
                                "facebook"
                            ]
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
...
        "/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "get the authenticated user's favorite books",
                "operationId": "33b5437261624b01bbac1eaef6f5871c",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/saves": {
            "get": {
                "tags": [
                    "Saves"
                ],
                "summary": "get the authenticated user's saved books",
                "operationId": "fc678fb9e94e7e4b2ae32e35571c51ae",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "BookResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ISBN": {
                        "type": "string",
                        "example": "9781234567897"
                    },
                    "title": {
                        "type": "string",
                        "example": "Clean Architecture"
                    },
                    "description": {
                        "type": "string",
                        "example": "A book about clean architecture"
                    },
                    "published_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}